implemented training utility
The objective maximizes the probability of making the correct choice. At first I had tried to use cross-entropy loss, but then I realized that's not what I want: cross-entropy tries to predict the posterior distribution as closely as possible. What I want is a strategy, and that strategy might (and, I believe, does) putting more probability on a given choice than you actually believe the choice is worth.